node-client/verify_system.py
"""
OBLIVION System Verification
Demonstrates the decentralized ML marketplace flow
"""
import os
from dotenv import load_dotenv
from web3 import Web3
from web3.middleware import ExtraDataToPOAMiddleware
from eth_account import Account

load_dotenv()

print("=" * 60)
print("   OBLIVION DECENTRALIZED SYSTEM VERIFICATION")
print("=" * 60)

# Connect to blockchain
rpc_url = os.getenv('RPC_URL') or os.getenv('POLYGON_RPC_URL')
contract_address = os.getenv('CONTRACT_ADDRESS')
private_key = os.getenv('PRIVATE_KEY')

w3 = Web3(Web3.HTTPProvider(rpc_url))
w3.middleware_onion.inject(ExtraDataToPOAMiddleware, layer=0)

account = Account.from_key(private_key)
print(f"\n✅ Connected to Polygon Amoy")
print(f"   Block: {w3.eth.block_number}")
print(f"   Contract: {contract_address}")
print(f"   Wallet: {account.address}")
print(f"   Balance: {w3.eth.get_balance(account.address) / 10**18:.4f} MATIC")

# Load contract
ABI = [
    {"inputs": [], "name": "getJobCount", "outputs": [{"type": "uint256"}], "stateMutability": "view", "type": "function"},
    {"inputs": [{"type": "uint256"}], "name": "getJob", "outputs": [{"type": "address"},{"type": "address"},{"type": "uint256"},{"type": "uint256"},{"type": "string"},{"type": "string"},{"type": "string"},{"type": "uint256"}], "stateMutability": "view", "type": "function"},
    {"inputs": [], "name": "getStats", "outputs": [{"type": "uint256"},{"type": "uint256"},{"type": "uint256"},{"type": "uint256"}], "stateMutability": "view", "type": "function"},
    {"inputs": [{"type": "address"}], "name": "getWorker", "outputs": [{"type": "uint256"},{"type": "uint256"},{"type": "uint256"},{"type": "bool"},{"type": "string"}], "stateMutability": "view", "type": "function"},
    {"inputs": [], "name": "MIN_STAKE", "outputs": [{"type": "uint256"}], "stateMutability": "view", "type": "function"},
]

contract = w3.eth.contract(address=contract_address, abi=ABI)

# Get stats
stats = contract.functions.getStats().call()
print(f"\n📊 ON-CHAIN STATISTICS:")
print(f"   Total Jobs: {stats[0]}")
print(f"   Pending Jobs: {stats[1]}")
print(f"   Completed Jobs: {stats[2]}")
print(f"   Active Workers: {stats[3]}")

# Get worker info
worker = contract.functions.getWorker(account.address).call()
print(f"\n👷 WORKER STATUS:")
print(f"   Stake: {worker[0] / 10**18:.4f} MATIC")
print(f"   Completed Jobs: {worker[1]}")
print(f"   Reputation: {worker[2]}")
print(f"   Active: {worker[3]}")
print(f"   Node ID: {worker[4]}")

# List jobs
job_count = contract.functions.getJobCount().call()
print(f"\n📋 JOBS ON-CHAIN ({job_count} total):")

STATUS_NAMES = ['Pending', 'Processing', 'Completed', 'Cancelled', 'Slashed']

for i in range(job_count):
    job = contract.functions.getJob(i).call()
    # job: (requester, worker, reward, status, scriptHash, dataHash, modelHash, createdAt)
    status = STATUS_NAMES[job[3]] if job[3] < len(STATUS_NAMES) else 'Unknown'
    print(f"\n   Job #{i}:")
    print(f"     Status: {status}")
    print(f"     Reward: {job[2] / 10**18:.4f} MATIC")
    print(f"     Script: {job[4][:40]}...")
    print(f"     Data: {job[5][:40]}...")
    if job[6]:
        print(f"     Model: {job[6][:40]}...")
    print(f"     Requester: {job[0][:20]}...")
    if job[1] != '0x0000000000000000000000000000000000000000':
        print(f"     Worker: {job[1][:20]}...")

print("\n" + "=" * 60)
print("   ARCHITECTURE VERIFICATION")
print("=" * 60)

print("""
✅ Smart Contract: OblivionManagerSimple deployed on Polygon Amoy
   - Job creation with ETH rewards
   - Worker registration with staking
   - Job claiming and completion
   - On-chain statistics

✅ IPFS Integration: For file storage
   - Training scripts (scriptHash)
   - Datasets (dataHash)
   - Trained models (modelHash)

✅ Worker Features:
   - Fair job distribution
   - Differential privacy (ε=1.0)
   - Quality verification
   - Stake-based incentives

✅ No External Database Required!
   - All coordination on-chain
   - IPFS for file storage
   - Fully decentralized
""")

# Min stake
min_stake = contract.functions.MIN_STAKE().call()
print(f"📌 Minimum Stake: {min_stake / 10**18:.4f} MATIC")

print("\n" + "=" * 60)
print("✅ DECENTRALIZED SYSTEM OPERATIONAL!")
print("=" * 60)
