node-client/requirements.txt
supabase>=2.0.0
web3>=6.0.0
ezkl>=7.0.0
numpy>=1.24.0
onnx>=1.14.0
python-dotenv>=1.0.0
requests>=2.28.0
torch>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
pandas>=2.0.0
